+++ src/egon/data/datasets/gas_prod.py
@@ -19,7 +19,7 @@ class GasProduction(Dataset):
      def __init__(self, dependencies):
          super().__init__(
              name="GasProduction",
-             version="0.0.3",
+             version="0.0.4",
              dependencies=dependencies,
              tasks=(import_gas_generators),
          )
@@ -151,7 +151,7 @@ def load_biogas_generators():
             AND ST_Contains(ST_Transform(vg.geometry,4326), egon_biogas_generator.geom)'''
 
         biogas_generators_list = gpd.GeoDataFrame.from_postgis(sql, con=engine, geom_col="geom", crs=4326)
-        biogas_generators_list = biogas_generators_list.drop(columns=['gid', 'bez', 'area_ha', 'geometry'])
+        biogas_generators_list = biogas_generators_list.drop(columns=['id', 'bez', 'area_ha', 'geometry'])
         db.execute_sql(
             """
               DROP TABLE IF EXISTS grid.egon_biogas_generator CASCADE;